backend/requirements.txt
# ──────────────────────────────────────────────
# Ultra Doc-Intelligence — Backend Dependencies
# ──────────────────────────────────────────────

# ── Core Framework ──
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# ── Data Validation & Config ──
pydantic>=2.10.0
python-dotenv>=1.0.1

# ── Document Parsing ──
pypdf>=4.0.0
python-docx>=1.0.0

# ── AI / LLM / Embeddings ──
huggingface-hub>=0.27.0
langchain>=0.3.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
sentence-transformers>=3.0.0

# ── Vector Store ──
faiss-cpu>=1.7.4

# ── Numerical / Scientific ──
numpy>=1.24.0

# ── Encryption / Security ──
cryptography>=42.0.0

# ── HTTP & Networking ──
requests>=2.31.0
httpx>=0.27.0
aiofiles>=24.1.0

# ── Logging & Monitoring ──
loguru>=0.7.0

# ── Testing (dev) ──
pytest>=8.0.0
pytest-asyncio>=0.24.0
